refactor and modify procfile
