+++ suslab/__init__.py
@@ -1,7 +1,7 @@
 from flask import Flask, render_template, Blueprint
 from flask_sqlalchemy import SQLAlchemy
 from flask_wtf.csrf import CSRFProtect
-from things_library.products.controllers import library
+from suslab.library.controllers import library
 
 app = Flask(__name__)
 app.config.from_object('config')
@@ -20,3 +20,8 @@ def index():
 @app.errorhandler(404)
 def not_found(error):
     return render_template('404.html'), 404
+
+
+def get_app():
+    '''Return the app by initializing components'''
+    return app
