+++ suslab/static/js/plugins.js
